requirements.txt
imutils==0.5.4
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
